refactor: Move constants values to separate file
