+++ lib/src/widget/built_in/built_in_style.dart
@@ -3,34 +3,6 @@ import 'package:toastification/src/helper/toast_helper.dart';
 import 'package:toastification/toastification.dart';
 
 /// default style for info toastification
-const infoColor = Color(0xFF47AFFF);
-
-/// default color for success toastification
-const successColor = Color(0xFF32BC32);
-
-/// default color for warning toastification
-const warningColor = Color(0xFFFFB600);
-
-/// default color for error toastification
-const errorColor = Color(0xFFFF3A30);
-
-const lowModeShadow = [
-  BoxShadow(
-    color: Color(0x07000000),
-    blurRadius: 16,
-    offset: Offset(0, 16),
-    spreadRadius: 0,
-  )
-];
-
-const highModeShadow = [
-  BoxShadow(
-    color: Color(0x14000000),
-    blurRadius: 30,
-    offset: Offset(0, 20),
-    spreadRadius: 0,
-  )
-];
 
 /// Base abstract class for built-in styles
 abstract class BuiltInStyle {
